requirements.txt
undetected-chromedriver
selenium
pandas
numpy
torch
scikit-learn
joblib
matplotlib
python-telegram-bot
groq
